Improving TextEdit text position
When text is bigger than the TextEdit, left part of the text is hidden
so that user sees only the end of his entry. It avoid the text to get
out of the TextEdit.
But, there is a problem whith non monospaced fonts, as number of
characters to show is calculated from the size of one letter. To fix, if
it is possible to do so.