Fix the TextEdit problem with non monospaces fonts
There was a problem with non monospaces fonts : as the number of
characters to show was calculated from the size of one letter.
Now, it's not based on a number of words, but in a loop which adds
letter until size reach the TextEdit one. It's heavier (maybe a little
too).