Chekbox and TextEdit improvement
In the TextEdit, position of the cursor exists now. It allow the user to
add a char everywhere in the TextEdit, and to delete chars via the Del
key (or backspace too). But, for now, move of the cursor is not fully
managed : when the cursor reach the left or right border, text is not
moved as it should be. TODO